backend/requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-ollama
chromadb
fastapi
uvicorn
pypdf
python-dotenv